[REFACTOR/#402] 바텀시트 위치 옮기기
